+++ assets/license-header.js
@@ -3,6 +3,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // --
-// Copyright 2016-2018 Andi Dittrich <https://andidittrich.de>
+// Copyright 2016-2019 Andi Dittrich <https://andidittrich.de>
 // ----------------------------------------------------------------------