+++ src/ui/views/codegroup.jsx
@@ -52,6 +52,11 @@ export function codegroup(dataset){
         cssClasses.push('enlighter-overflow-scroll');
     }
 
+    // additional css classes set ?
+    if (options.cssClasses.length > 0){
+        cssClasses.push(...options.cssClasses);
+    }
+
     // utility function to toggle raw code
     function toggleContainer(){
         React.toggleClass(wrapper, 'enlighter-show-rawcode');
@@ -98,7 +103,7 @@ export function codegroup(dataset){
 
     // dbclick event ?
     if (options.rawcodeDbclick){
-        wrapper.on('dbclick', toggleContainer);
+        wrapper.on('dblclick', toggleContainer);
     }
 
     return wrapper;